requirements_pip.txt
aiobotocore @ file:///croot/aiobotocore_1680004275080/work
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
asttokens==2.2.1
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs==22.2.0
awscli==1.27.103
backcall==0.2.0
boto3==1.26.103
botocore==1.29.105
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==2.0.12
colorama==0.4.4
comm==0.1.2
cryptography @ file:///croot/cryptography_1677533068310/work
debugpy==1.6.6
distro==1.8.0
docutils==0.16
executing==1.2.0
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
frozenlist @ file:///croot/frozenlist_1670004507010/work
fsspec @ file:///croot/fsspec_1679418786824/work
greenlet==2.0.2
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.20.2
ipython==8.9.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.2
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.0
requests==2.28.2
rsa==4.7.2
s3fs==0.4.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
tabula-py==2.7.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3 @ file:///croot/urllib3_1680254681959/work
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.8.2
